+++ hardware/vsim/start_sim_with_dpi.sh
@@ -1,5 +1,5 @@
 #!/usr/bin/env bash
 set -e
 
-SIMSCRIPT=run_pulp_runtime.tcl
+SIMSCRIPT=run_with_dpi.tcl
 source sim_common_post.sh